feat(napi): bridge the sort helpers, the first module with no instance
Four #[napi] free functions and one helper, plus the shims that put upstream's
two-file export shape back together out of the addon's flat namespace
(DESIGN.md 2.3, Problem 2). tests/bridge/sort.js is where that assembly lives;
sort/insertion.js and sort/quick.js are cut from it so nothing is decorative
and gate 3 has a shim named after the unit.

test/sort.js is green: 13 passing, file unmodified.

Two things the bridge has to carry rather than decide:

* In place means in place. Array<'env> and the typed arrays both carry their
  original napi_value through ToNapiValue, so the returned object IS the
  argument. test/sort.js never checks that -- it only inspects the return
  value -- so a port that sorted a copy would pass all thirteen assertions and
  break every in-tree caller. tests/boundary/sort.js asserts it.
* indices() takes an f64 because upstream's two uses of the argument coerce
  DIFFERENTLY, and the difference is observable. getPointerArray compares
  length - 1 as a double; the TypedArray constructor applies ToIndex, which
  truncates. So indices(256.5) is a Uint16Array of 256 elements -- a width
  wider than 256 elements need. Confirmed against Node 24.18.1. The first
  draft took a usize, truncated first, and would have produced Uint8Array;
  the boundary spec is what caught it.

Stated divergences, all refusals with messages that name the limit: non-numeric
elements inside the window (upstream compares through valueOf, which is T2),
and lo/hi outside 0..=length (upstream reads undefined and writes into holes).

tests/boundary/sort.js covers what neither the original suite nor cargo test
can reach: object identity, mutation of the caller's typed array, every window
of a fixed array against vendored upstream, the Uint16Array width upstream
never constructs, and each refusal.

+++ crates/mnemonist-core/src/sort/insertion.rs
@@ -101,7 +101,7 @@ mod tests {
     const DATA: [i32; 11] = [2, 7, 1, 5, 8, 9, 1, -3, 3, 18, 6];
 
     fn indices_of(len: usize) -> PointerVec {
-        crate::utils::typed_arrays::indices(len).expect("len fits a pointer array")
+        crate::utils::typed_arrays::indices(len as f64).expect("len fits a pointer array")
     }
 
     fn as_vec(indices: &PointerVec) -> Vec<u32> {
